[refactor] BookingViewController의 TimeOptionCollectionView를 TimeOptionView로 이관
